Add files via upload
The login And Registration System project in C++ involves mainly the user registration process. User Credentials like usernames and passwords are asked from the user. If the registration of the user is successful then with the given credentials a file will be created of a particular user in the database.